+++ src/sentry/api/endpoints/group_notes.py
@@ -8,7 +8,8 @@ from sentry.api.bases.group import GroupEndpoint
 from sentry.api.serializers import serialize
 from sentry.api.serializers.rest_framework.group_notes import NoteSerializer
 from sentry.api.serializers.rest_framework.mentions import extract_user_ids_from_mentions
-from sentry.models import Activity, GroupSubscription, GroupSubscriptionReason
+from sentry.models import Activity, GroupSubscription
+from sentry.notifications.types import GroupSubscriptionReason
 from sentry.utils.functional import extract_lazy_object
