+++ src/sentry/mail/activity/new_processing_issues.py
@@ -1,5 +1,6 @@
-from sentry.models import EventError, GroupSubscriptionReason, NotificationSetting
+from sentry.models import EventError, NotificationSetting
 from sentry.models.integration import ExternalProviders
+from sentry.notifications.types import GroupSubscriptionReason
 from sentry.utils.http import absolute_uri
 
 from .base import ActivityEmail
