+++ src/sentry/models/groupassignee.py
@@ -7,6 +7,7 @@ from django.utils import timezone
 
 from sentry.db.models import BaseManager, FlexibleForeignKey, Model, sane_repr
 from sentry.models.activity import Activity
+from sentry.notifications.types import GroupSubscriptionReason
 from sentry.signals import issue_assigned
 from sentry.utils import metrics
 
@@ -125,7 +126,7 @@ def sync_group_assignee_outbound(group, user_id, assign=True):
 class GroupAssigneeManager(BaseManager):
     def assign(self, group, assigned_to, acting_user=None):
         from sentry import features
-        from sentry.models import GroupSubscription, GroupSubscriptionReason, Team, User
+        from sentry.models import GroupSubscription, Team, User
 
         GroupSubscription.objects.subscribe_actor(
             group=group, actor=assigned_to, reason=GroupSubscriptionReason.assigned
