+++ src/sentry/web/frontend/debug/debug_new_release_email.py
@@ -3,16 +3,8 @@ import datetime
 import pytz
 from django.views.generic import View
 
-from sentry.models import (
-    Commit,
-    CommitAuthor,
-    Deploy,
-    GroupSubscriptionReason,
-    Organization,
-    Project,
-    Release,
-    User,
-)
+from sentry.models import Commit, CommitAuthor, Deploy, Organization, Project, Release, User
+from sentry.notifications.types import GroupSubscriptionReason
 from sentry.utils.compat import zip
 from sentry.utils.http import absolute_uri
